A bot to write Students detail from excel in notepad.
This bot is made such that the details of every student(Roll No, Name, Mobile number) is written on separate notepads.The number of notepads created is equal to number of rows in excel file.Then, the details of every student is written in their respected notepads and the values are comma separated. Each notepad is saved in the name of the student by the bot.