A bot to validate Adhar and PAN card number.
This bot asks for the adhar card number, if the pattern is valid it says that the given number is valid and saves the given adhar number in notepad with the name as "Adhar" but if the given number is not valid it shows the user that the input was invalid. Then,the bot asks for the PAN card number and undergoes the same procedure as it went for adhar card and the valid PAN number is saved in notepad with file name"PAN".Here to validate the pattern of the card details, REGEX is used.