A bot to work on QR code for clients.
This bot generates QR or Quick Response code for each client in the excel mentioned based on their name and ID and then saves it in the downloads folder in png format and then mails it to the client. Here the bot reads the excel file through message box at the beginning and then the message that confirms the mail forwards is also showed to the user using message box.