A guess game in RPA.
The bot will generate a number between 1-10 and will ask the user to guess it correctly, through input dialog box. If the user is either guessing a greater or smaller number, then the message box tells that the guessed number is greater or smaller, respectively. The bot runs until the guessed number matches the number generated by the bot. When it  matches, the message box confirms it.The work flow used here is State machine.