A bot to work on custom input form.
A custom input form has been created using HTML. The path of the form is being provided in the workflow. The details are asked to the user dynamically according to the form when the bot is run. The details are being saved in the excel mentioned and mailed to the mail ID given.